PM-13 fix based on comments

+++ src/main/java/com/pik01/pharmaciesmanager/app/medicine/MedicineController.java
@@ -26,7 +26,7 @@ public class MedicineController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public void deleteMed(@PathVariable("id") int id ) {
+    public void deleteMed(@PathVariable("id") long id ) {
         Long idLong = new Long(id);
         medicineService.delete(idLong);
     }
